requirements.txt
# Web Framework
Flask>=2.1.1
Werkzeug>=2.1.1
gunicorn>=20.1.0; platform_system != 'Windows'  # Production WSGI server (Unix only)
waitress>=2.1.0; platform_system == 'Windows'    # Alternative for Windows

# LlamaIndex Core & Integrations
llama-index-core>=0.10.0
llama-index-llms-siliconflow>=0.1.0
llama-index-embeddings-siliconflow>=0.1.0
llama-index-embeddings-huggingface
llama-index-graph-stores-neo4j
llama-index-readers-file

# Database
neo4j>=5.0.0

# Cloud Storage
cos-python-sdk-v5>=1.9.0

# AI & ML
torch>=2.0.0
transformers>=4.30.0
sentence-transformers
numpy>=2.2.6
pandas>=2.0.0
pillow>=10.0.0
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
tqdm>=4.66.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
openai>=1.0.0

# Text Processing
nltk>=3.8.1
docx2txt>=0.8
pypdf>=3.0.0
